refactor: 컨벤션에 맞게 패키지 및 클래스명 수정
