--- src/main/java/com/chungnamthon/cheonon/Auth/Dto/ApiResponse.java
@@ -1,32 +1,0 @@
-package com.chungnamthon.cheonon.Auth.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-@Builder
-public class ApiResponse<T> {
-    private String message;
-    private LocalDateTime timestamp;
-    private T data;
-
-    public static <T> ApiResponse<T> success(String message, T data) {
-        return ApiResponse.<T>builder()
-                .message(message)
-                .timestamp(LocalDateTime.now())
-                .data(data)
-                .build();
-    }
-
-    public static <T> ApiResponse<T> error(String message) {
-        return ApiResponse.<T>builder()
-                .message(message)
-                .timestamp(LocalDateTime.now())
-                .data(null)
-                .build();
-    }
-}
