--- src/main/java/com/chungnamthon/cheonon/Auth/Dto/RefreshTokenRequestDto.java
@@ -1,10 +1,0 @@
-package com.chungnamthon.cheonon.Auth.Dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class RefreshTokenRequestDto {
-    private String refreshToken;
-}
