--- src/main/java/com/chungnamthon/cheonon/Auth/Dto/TokenResponseDto.java
@@ -1,11 +1,0 @@
-package com.chungnamthon.cheonon.Auth.Dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TokenResponseDto {
-    private String accessToken;
-    private String refreshToken; // 필요 시 null일 수 있음
-}
