+++ src/main/java/com/chungnamthon/cheonon/auth/dto/request/RefreshTokenRequest.java
@@ -1,0 +1,10 @@
+package com.chungnamthon.cheonon.auth.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class RefreshTokenRequest {
+    private String refreshToken;
+}
