+++ src/main/java/com/chungnamthon/cheonon/auth/dto/response/ApiResponse.java
@@ -1,0 +1,32 @@
+package com.chungnamthon.cheonon.auth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+@Builder
+public class ApiResponse<T> {
+    private String message;
+    private LocalDateTime timestamp;
+    private T data;
+
+    public static <T> ApiResponse<T> success(String message, T data) {
+        return ApiResponse.<T>builder()
+                .message(message)
+                .timestamp(LocalDateTime.now())
+                .data(data)
+                .build();
+    }
+
+    public static <T> ApiResponse<T> error(String message) {
+        return ApiResponse.<T>builder()
+                .message(message)
+                .timestamp(LocalDateTime.now())
+                .data(null)
+                .build();
+    }
+}
