+++ src/main/java/com/chungnamthon/cheonon/auth/dto/response/TokenResponse.java
@@ -1,0 +1,11 @@
+package com.chungnamthon.cheonon.auth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenResponse {
+    private String accessToken;
+    private String refreshToken; // 필요 시 null일 수 있음
+}
