--- src/main/java/com/chungnamthon/cheonon/common/GlobalExceptionHandler.java
@@ -1,26 +1,0 @@
-package com.chungnamthon.cheonon.common;
-
-import com.chungnamthon.cheonon.Auth.Dto.ApiResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    // IllegalArgumentException 처리
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<ApiResponse<Void>> handleIllegalArgument(IllegalArgumentException ex) {
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ApiResponse.error(ex.getMessage()));
-    }
-
-    // 그 외 Exception 처리
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ApiResponse<Void>> handleException(Exception ex) {
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ApiResponse.error("서버 오류가 발생했습니다."));
-    }
-}
