+++ src/main/java/com/chungnamthon/cheonon/global/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.chungnamthon.cheonon.global.config;
 
-import com.chungnamthon.cheonon.Auth.Jwt.JwtFilter;
+import com.chungnamthon.cheonon.auth.jwt.JwtFilter;
 import com.chungnamthon.cheonon.global.security.CustomAuthenticationEntryPoint;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
