+++ src/main/java/com/chungnamthon/cheonon/global/exception/error/AuthenticationError.java
@@ -1,9 +1,11 @@
 package com.chungnamthon.cheonon.global.exception.error;
 
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @Getter
+@RequiredArgsConstructor
 public enum AuthenticationError implements BaseError {
 
     INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "유효하지 않은 토큰입니다."),
@@ -11,9 +13,4 @@ public enum AuthenticationError implements BaseError {
 
     private final HttpStatus httpStatus;
     private final String message;
-
-    AuthenticationError(HttpStatus httpStatus, String message) {
-        this.httpStatus = httpStatus;
-        this.message = message;
-    }
 }
