--- src/main/java/com/chungnamthon/cheonon/localmerchant/Dto/LocalMerchantDto.java
@@ -1,20 +1,0 @@
-package com.chungnamthon.cheonon.localmerchant.Dto;
-
-import lombok.*;
-
-import java.math.BigInteger;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class LocalMerchantDto {
-    private BigInteger merchantSeq;
-    private String name;
-    private String category;
-    private String address;
-    private String tel;
-    private Double latitude;
-    private Double longitude;
-}
