+++ src/main/java/com/chungnamthon/cheonon/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.chungnamthon.cheonon.user.controller;
+
+public class UserController {
+}
