+++ src/main/java/com/chungnamthon/cheonon/user/dto/UserDto.java
@@ -1,0 +1,15 @@
+package com.chungnamthon.cheonon.user.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UserDto {
+    private Long userId;
+    private String email;
+    private String image;
+    private String nickname;
+}
