+++ src/main/java/com/chungnamthon/cheonon/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.chungnamthon.cheonon.user.repository;
+
+import com.chungnamthon.cheonon.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
