--- src/main/java/com/chungnamthon/cheonon/users/Controller/Controller.java
@@ -1,4 +1,0 @@
-package com.chungnamthon.cheonon.users.Controller;
-
-public class Controller {
-}
